refactor: extract shell completion scripts
