+++ packages/cli/src/completions/phantom-fish.ts
@@ -1,0 +1,107 @@
+export const FISH_COMPLETION_SCRIPT = `# Fish completion for phantom
+# Load with: phantom completion fish | source
+
+function __phantom_list_worktrees
+    phantom list --names 2>/dev/null
+end
+
+function __phantom_using_command
+    set -l cmd (commandline -opc)
+    set -l cmd_count (count $cmd)
+    if test $cmd_count -eq 1
+        # No subcommand yet, so any command can be used
+        if test (count $argv) -eq 0
+            return 0
+        else
+            return 1
+        end
+    else if test $cmd_count -ge 2
+        # Check if we're in the context of a specific command
+        if test (count $argv) -gt 0 -a "$argv[1]" = "$cmd[2]"
+            return 0
+        end
+    end
+    return 1
+end
+
+# Disable file completion for phantom
+complete -c phantom -f
+
+# Main commands
+complete -c phantom -n "__phantom_using_command" -a "create" -d "Create a new Git worktree (phantom)"
+complete -c phantom -n "__phantom_using_command" -a "attach" -d "Attach to an existing branch by creating a new worktree"
+complete -c phantom -n "__phantom_using_command" -a "list" -d "List all Git worktrees (phantoms)"
+complete -c phantom -n "__phantom_using_command" -a "where" -d "Output the filesystem path of a specific worktree"
+complete -c phantom -n "__phantom_using_command" -a "delete" -d "Delete a Git worktree (phantom)"
+complete -c phantom -n "__phantom_using_command" -a "exec" -d "Execute a command in a worktree directory"
+complete -c phantom -n "__phantom_using_command" -a "review" -d "Review changes in a worktree with a local PR review interface (experimental)"
+complete -c phantom -n "__phantom_using_command" -a "shell" -d "Open an interactive shell in a worktree directory"
+complete -c phantom -n "__phantom_using_command" -a "github" -d "GitHub integration commands"
+complete -c phantom -n "__phantom_using_command" -a "gh" -d "GitHub integration commands (alias)"
+complete -c phantom -n "__phantom_using_command" -a "version" -d "Display phantom version information"
+complete -c phantom -n "__phantom_using_command" -a "completion" -d "Generate shell completion scripts"
+complete -c phantom -n "__phantom_using_command" -a "mcp" -d "Manage Model Context Protocol (MCP) server"
+
+# Global options
+complete -c phantom -l help -d "Show help (-h)"
+complete -c phantom -l version -d "Show version (-v)"
+
+# create command options
+complete -c phantom -n "__phantom_using_command create" -l shell -d "Open an interactive shell in the new worktree after creation (-s)"
+complete -c phantom -n "__phantom_using_command create" -l exec -d "Execute a command in the new worktree after creation (-x)" -x
+complete -c phantom -n "__phantom_using_command create" -l tmux -d "Open the worktree in a new tmux window (-t)"
+complete -c phantom -n "__phantom_using_command create" -l tmux-vertical -d "Open the worktree in a vertical tmux pane"
+complete -c phantom -n "__phantom_using_command create" -l tmux-horizontal -d "Open the worktree in a horizontal tmux pane"
+complete -c phantom -n "__phantom_using_command create" -l copy-file -d "Copy specified files from the current worktree" -r
+complete -c phantom -n "__phantom_using_command create" -l base -d "Branch or commit to create the new worktree from (defaults to HEAD)" -x
+
+# attach command options
+complete -c phantom -n "__phantom_using_command attach" -l shell -d "Open an interactive shell in the worktree after attaching (-s)"
+complete -c phantom -n "__phantom_using_command attach" -l exec -d "Execute a command in the worktree after attaching (-x)" -x
+
+# list command options
+complete -c phantom -n "__phantom_using_command list" -l fzf -d "Use fzf for interactive selection"
+complete -c phantom -n "__phantom_using_command list" -l names -d "Output only phantom names (for scripts and completion)"
+
+# where command options
+complete -c phantom -n "__phantom_using_command where" -l fzf -d "Use fzf for interactive selection"
+complete -c phantom -n "__phantom_using_command where" -a "(__phantom_list_worktrees)"
+
+# delete command options
+complete -c phantom -n "__phantom_using_command delete" -l force -d "Force deletion even if worktree has uncommitted changes (-f)"
+complete -c phantom -n "__phantom_using_command delete" -l current -d "Delete the current worktree"
+complete -c phantom -n "__phantom_using_command delete" -l fzf -d "Use fzf for interactive selection"
+complete -c phantom -n "__phantom_using_command delete" -a "(__phantom_list_worktrees)"
+
+# exec command options
+complete -c phantom -n "__phantom_using_command exec" -l fzf -d "Use fzf for interactive selection"
+complete -c phantom -n "__phantom_using_command exec" -l tmux -d "Execute command in new tmux window (-t)"
+complete -c phantom -n "__phantom_using_command exec" -l tmux-vertical -d "Execute command in vertical split pane"
+complete -c phantom -n "__phantom_using_command exec" -l tmux-horizontal -d "Execute command in horizontal split pane"
+complete -c phantom -n "__phantom_using_command exec" -a "(__phantom_list_worktrees)"
+
+# review command options
+complete -c phantom -n "__phantom_using_command review" -l fzf -d "Use fzf for interactive selection"
+complete -c phantom -n "__phantom_using_command review" -l base -d "Base reference for comparison" -x
+complete -c phantom -n "__phantom_using_command review" -a "(__phantom_list_worktrees)"
+
+# shell command options
+complete -c phantom -n "__phantom_using_command shell" -l fzf -d "Use fzf for interactive selection"
+complete -c phantom -n "__phantom_using_command shell" -l tmux -d "Open shell in new tmux window (-t)"
+complete -c phantom -n "__phantom_using_command shell" -l tmux-vertical -d "Open shell in vertical split pane"
+complete -c phantom -n "__phantom_using_command shell" -l tmux-horizontal -d "Open shell in horizontal split pane"
+complete -c phantom -n "__phantom_using_command shell" -a "(__phantom_list_worktrees)"
+
+# completion command - shell names
+complete -c phantom -n "__phantom_using_command completion" -a "fish zsh bash" -d "Shell type"
+
+# github command options
+complete -c phantom -n "__phantom_using_command github" -a "checkout" -d "Create a worktree for a GitHub PR or issue"
+complete -c phantom -n "__phantom_using_command gh" -a "checkout" -d "Create a worktree for a GitHub PR or issue"
+
+# github checkout command options
+complete -c phantom -n "__phantom_using_command github checkout" -l base -d "Base branch for new issue branches (issues only)" -x
+complete -c phantom -n "__phantom_using_command gh checkout" -l base -d "Base branch for new issue branches (issues only)" -x
+
+# mcp command options
+complete -c phantom -n "__phantom_using_command mcp" -a "serve" -d "Start MCP server"`;
